3.Airflow_Spark/requirements.txt
pandas
pandasql
sqlalchemy
yadisk

pyspark==3.1.2
apache-airflow-providers-apache-spark==2.0.0